# Conflicts:
#	pxr/imaging/lib/garch/textureHandle.cpp
#	pxr/imaging/lib/garch/textureRegistry.cpp
#	pxr/imaging/lib/hd/unitTestNullRenderDelegate.cpp
#	pxr/imaging/lib/hdSt/codeGen.cpp
#	pxr/imaging/lib/hdSt/dispatchBuffer.cpp
#	pxr/imaging/lib/hdSt/interleavedMemoryManager.h
#	pxr/imaging/lib/hdSt/material.cpp
#	pxr/imaging/lib/hdSt/shaders/terminals.glslfx
#	pxr/imaging/lib/hdSt/textureResource.cpp
#	pxr/imaging/lib/hdSt/vboMemoryManager.h
#	pxr/imaging/lib/hdSt/vboSimpleMemoryManager.h
#	pxr/usdImaging/lib/usdImagingGL/hydraMaterialAdapter.cpp

+++ pxr/imaging/lib/hd/unitTestNullRenderDelegate.cpp
@@ -114,7 +114,7 @@ public:
             return 0;
         }
 
-        virtual int GetNumElements() const override {
+        virtual size_t GetNumElements() const override {
             return _numElements;
         }
 
@@ -162,7 +162,7 @@ public:
 
     private:
         _BufferArray * _bufferArray;
-        int _numElements;
+        size_t _numElements;
     };
